first pass as perf, looks great but brush is out of sync
