+++ rollplay/app/fog_management/components/index.js
@@ -1,7 +1,6 @@
 /* Copyright (C) 2025 Matthew Davey */
 /* SPDX-License-Identifier: GPL-3.0-or-later */
 
-export { default as FogCanvasLayer } from './FogCanvasLayer';
 export { default as FogRegionStack } from './FogRegionStack';
 export { default as FogRegionLabels } from './FogRegionLabels';
 export { default as FogPaintControls } from './FogPaintControls';
